Added run_metrics test
